requirements.txt
datadog-checks-base==11.0.0
prometheus-client==0.7.1
PyYAML==5.3.1
requests==2.23.0
six==1.14.0
